Revert #1694 "[LiveComponent] set LiveArg value to null if empty string"
