+++ src/LiveComponent/src/Attribute/LiveArg.php
@@ -11,49 +11,44 @@
 
 namespace Symfony\UX\LiveComponent\Attribute;
 
-use Symfony\Component\HttpKernel\Attribute\ValueResolver;
-use Symfony\UX\LiveComponent\ValueResolver\LiveArgValueResolver;
-
-if (class_exists(ValueResolver::class)) {
-    /**
-     * An attribute to configure a LiveArg (custom argument passed to a LiveAction).
-     *
-     * @see https://symfony.com/bundles/ux-live-component/current/index.html#actions-arguments
-     *
-     * @author Tomas Norkūnas <norkunas.tom@gmail.com>
-     * @author Jannes Drijkoningen <jannesdrijkoningen@gmail.com>
-     */
-    #[\Attribute(\Attribute::TARGET_PARAMETER)]
-    final class LiveArg extends ValueResolver
-    {
-        public function __construct(
-            /**
-             * @param string|null $name The name of the argument received by the LiveAction
-             */
-            public ?string $name = null,
-            bool $disabled = false,
-            string $resolver = LiveArgValueResolver::class,
-        ) {
-            parent::__construct($resolver, $disabled);
-        }
+/**
+ * An attribute to configure a LiveArg (custom argument passed to a LiveAction).
+ *
+ * @see https://symfony.com/bundles/ux-live-component/current/index.html#actions-arguments
+ *
+ * @author Tomas Norkūnas <norkunas.tom@gmail.com>
+ */
+#[\Attribute(\Attribute::TARGET_PARAMETER)]
+final class LiveArg
+{
+    public function __construct(
+        /**
+         * @param string|null $name The name of the argument received by the LiveAction
+         */
+        public ?string $name = null,
+    ) {
     }
-} else {
+
     /**
-     * An attribute to configure a LiveArg (custom argument passed to a LiveAction).
+     * @internal
      *
-     * @see https://symfony.com/bundles/ux-live-component/current/index.html#actions-arguments
-     *
-     * @author Tomas Norkūnas <norkunas.tom@gmail.com>
+     * @return array<string, string>
      */
-    #[\Attribute(\Attribute::TARGET_PARAMETER)]
-    final class LiveArg
+    public static function liveArgs(object $component, string $action): array
     {
-        public function __construct(
-            /**
-             * @param string|null $name The name of the argument received by the LiveAction
-             */
-            public ?string $name = null,
-        ) {
+        $method = new \ReflectionMethod($component, $action);
+        $liveArgs = [];
+
+        foreach ($method->getParameters() as $parameter) {
+            foreach ($parameter->getAttributes(self::class) as $liveArg) {
+                /** @var LiveArg $attr */
+                $attr = $liveArg->newInstance();
+                $parameterName = $parameter->getName();
+
+                $liveArgs[$parameterName] = $attr->name ?? $parameterName;
+            }
         }
+
+        return $liveArgs;
     }
 }
