+++ src/LiveComponent/tests/Fixtures/templates/components/component6.html.twig
@@ -2,6 +2,4 @@
     Arg1: {{ this.called ? this.arg1 : 'not provided' }}
     Arg2: {{ this.called ? this.arg2 : 'not provided' }}
     Arg3: {{ this.called ? this.arg3 : 'not provided' }}
-    Arg4: {{ this.called ? (this.arg4 ?? 'null') : 'not provided' }}
-    Arg5: {{ this.called ? (this.arg5 is empty ? 'empty-string' : '') : 'not provided' }}
 </div>
